Create list describing parameters for drug efficacy and prophlyaxis

Create list describing parameters for drug efficacy and prophlyaxis

drug_create(
  prob_of_lpf = c(1, 0.97, 0.8, 0.55),
  barcode_res_pos = c(0, 1),
  prophylactic_pos = 1,
  dur_P = 25,
  dur_SPC = 5,
  drug_clearance_max_time = 60,
  prophylactic_probability = 1 - pgamma(seq(0, drug_clearance_max_time, 0.2), shape =
    16.8, rate = 16.8/17.9),
  prophylactic_resistant_probability = 1 - pgamma(seq(0, drug_clearance_max_time, 0.2),
    shape = 16.8, rate = 16.8/8.7)
)

Arguments

prob_of_lpf
Vector of probabilities for the chance of late parasitological failure (lpf). The vector gives the prob of lpf for each relevent barcode combination for a given drug. E.g. The default is:

c(1.0, 0.97, 0.80, 0.55)

The vector is length 4, and so 2 barcode positions change the prob of lpf. If the parasite is 0,0 then the prob of lpf is 0 (1-1), but if it was 0,1 it would be 0.2.

barcode_res_pos
Vector for which barcode positions correspond to which drug resistance mechanism. E.g. the default is:

c(0,1)

Resistance to drug is encoded at barcode position 0 and 1

prophylactic_pos
Vector for which barcode positions determine the impact of drug resistance in shortening the effective prophylactic period. E.g. the default is c(1), which shows that for the drug, the prophylactic position is encoded in barcode position 1.

dur_P
Duration of prophylaxis in days. Default = 25.

dur_SPC
Duration of slow parasite clearance. Default = 5.

drug_clearance_max_time
Maximum number of days to which to consider waning prophylaxis. Default = 60 days.

prophylactic_probability
Vector of changing probability of reinfection due to waning prophylaxis. The last element reflects the probability after drug_clearance_max_time and the first element is the probability at time = 0 days.

prophylactic_resistant_probability
Vector of changing probability of reinfection due to waning prophylaxis when challenged by a parasite that is resistant to the partner drug. The last element reflects the probability after drug_clearance_max_time and the first element is the probability at time = 0 days.