Create vector adaptation list

List for vector adaptations relating to oocyst success

vector_adaptation_list_create(
  vector_adaptation_loci,
  vector_adaptation_flag = FALSE,
  local_oocyst_advantage = 0.5,
  gametocyte_sterilisation_flag = FALSE,
  gametocyte_sterilisation = 0.5,
  oocyst_reduction_by_artemisinin = 0.2
)

Arguments

vector_adaptation_loci
Vector of integers detailing which loci in the barcode correspond to the vector adaptation phenotype

vector_adaptation_flag
Boolean are we doing vector adaptation.

local_oocyst_advantage
Numeric for probability that non adapated parasites will get through. Default = 0.5

gametocyte_sterilisation_flag
Boolean for whether we are doing gametocye sterilisation as a result from artemisinin. Default = FALSE

gametocyte_sterilisation
Numeric for theimpact of artemisinin on male gametocytes. Default = 0.5, which causes the probability that a wild type male gametocyte will be chosen to be in oocysts is halved.

oocyst_reduction_by_artemisinin
Numeric for reduction in oocyst under artemisinin drug pressure. default = 0.2