Create drug list

List for simulating drug usage for resistance/mft variables

drug_list_create(
  resistance_flag = FALSE,
  number_of_resistance_loci = 2,
  artemisinin_loci = c(0),
  cost_of_resistance = c(0.99, 0.99),
  absolute_fitness_cost_flag = FALSE,
  epistatic_logic = NULL,
  number_of_drugs = 1,
  drugs = list(),
  mft_flag = FALSE,
  temporal_cycling = -1,
  sequential_cycling = -1,
  sequential_update = 3,
  drug_choice = 0,
  partner_drug_ratios = rep(1/number_of_drugs, number_of_drugs)
)

Arguments

resistance_flag
Boolean are we simulating resistance

number_of_resistance_loci
Numeric for number of res. loci

artemisinin_loci
Numerics for barcode positions that confer artemisinin resistance

cost_of_resistance
Numeric vector for fitness cost of each resistance loci.

absolute_fitness_cost_flag
Boolean are we simulating fitness costs as absolute (i.e. impactive onward transmission chance) or relative (resistant strains have decreases chance of being onwardly transmitted in mixed infections). Default = FALSE.

epistatic_logic
Is there compensatory relationships. i.e. what loci need to be true for resistance cost to exist. Default of NULL means that this becomes seq_len(number_of_resistance_loci), i.e. only dependent on their own loci. (TODO: Change this to be a list of length norl)

number_of_drugs
Numeric for number of drugs used

drugs
List of drugs that are being used, with each list element being created by drug_create.

mft_flag
Boolean are we doing mft

temporal_cycling
Numeric for when in years a drug switch occurs

sequential_cycling
Numeric for what perc. treatment failure before switch

sequential_update
How long does it take in years for sequential to be implemented

drug_choice
What's the default drug choice to begin. Default = 0

partner_drug_ratios
Numeric vector for ratio of first line drugs used